refactor(PN-11969): wip on tests PG
